# level2/task2/docs/swagger.yaml
swagger: "2.0"
basePath: /api/v1
definitions:
...
  controller.ContractResponse:
    properties:
      address:
        example: 0x742d35Cc6634C0532925a3b844Bc454e4438f44e
        type: string
      contract_type:
        example: SimpleStorage
        type: string
      error:
        example: 交易失败
        type: string
      txHash:
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        type: string
    type: object
  controller.ErrorResponse:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: 服务器内部错误
        type: string
    type: object
...
  controller.SimpleAuctionBidRequest:
    properties:
      bid_amount:
        type: string
    required:
    - bid_amount
    type: object
  controller.SimpleAuctionBidResponse:
    properties:
      address:
        type: string
      bid_amount:
        type: string
      txHash:
        type: string
    type: object
  controller.SimpleStorageGetValueResponse:
    properties:
      address:
        type: string
      value:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: 提供以太坊区块链相关的API服务，包括账户管理、交易处理、合约部署等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Web3 区块链接口服务
  version: "1.0"
paths:
...
  /contract/simpleauction/bid:
    post:
      consumes:
      - application/json
      description: 向拍卖合约发起竞拍请求
      parameters:
      - description: 竞拍金额信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.SimpleAuctionBidRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SimpleAuctionBidResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 调用SimpleAuction合约进行竞拍
      tags:
      - 合约操作
  /contract/simpleauction/withdraw:
    post:
      consumes:
      - application/json
      description: 结束竞拍并允许中标者支付款项，非中标者提取资金
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ContractResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 调用SimpleAuction合约结束拍卖并提取资金
      tags:
      - 合约操作
...
  /contracts/SimpleStorage/get:
    get:
      consumes:
      - application/json
      description: 调用SimpleStorage合约的get方法获取当前存储的值
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SimpleStorageGetValueResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 获取SimpleStorage的值
      tags:
      - SimpleStorage
  /contracts/SimpleStorage/set:
    post:
      consumes:
      - application/json
      description: 调用SimpleStorage合约的set方法设置存储值
      parameters:
      - description: 要存储的整数值
        in: body
        name: value
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ContractResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 设置SimpleStorage值
      tags:
      - SimpleStorage
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey